==> STEP 3 - cleaning data.ipynb <==
import requests
import pandas as pd
import numpy as np

data = pd.read_csv("data_dirty.csv")
data.head()

def to_float(x):
    if "," in str(x):
        x = x.replace(",","")
    return round(float(x),2)

# dropping duplicates
data = data.drop_duplicates()

# fill gaps in coal because it had only 1 price input per month
data["coal_monthly"] = data["coal_monthly"].fillna(method = "ffill")

# dropping palladium because it has just too many NaNs, I think I can go without that column
try:
    data = data.drop(["palladium"], axis = 1)
except:
    pass

data_prices = data.drop(["date", "year"], axis = 1)
data_time = data[["date","year"]].copy()
data_prices = data_prices.fillna(value = np.nan)
data_prices["gold"] = data_prices["gold"].astype(float)
data_prices["silver"] = data_prices["silver"].astype(float)
remove_comma = ["dowjones_industrial","dowjones_transport","nasdaq","snp"]
for col in remove_comma:
    data_prices[col] = data_prices[col].apply(to_float)
data_prices = data_prices.round(2)
data = pd.concat([data_time, data_prices], axis = 1)

# checkin NaNs
data.isna().sum()

drop some nan in gold column

# drop some NaNs in gold column 
data = data[~(data["date"] < "1992-02-01")]
data["gold"].isna().sum()

drop some nan in platinum, southern_copper, rio_tinto, royal_gold, newmont, agnico_mines

data = data[~(data["date"] < "1997-10-01")]

fill the rest of NaN with interpolate function

data = data.interpolate()
data.isna().sum()

data.to_csv("data_clean.csv")
